+++ spec/acfs/runner_spec.rb
@@ -37,6 +37,7 @@ describe ::Acfs::Runner do
   let(:adapter) { ::NullAdapter.new }
   let(:runner) { ::Acfs::Runner.new adapter }
   let(:collector) { NotificationCollector.new }
+  let(:collector2) { NotificationCollector.new }
 
   after do
     ::ActiveSupport::Notifications.notifier = \
@@ -46,12 +47,14 @@ describe ::Acfs::Runner do
   describe '#instrumentation' do
     before do
       ::ActiveSupport::Notifications.subscribe /^acfs\.runner/, collector
+      ::ActiveSupport::Notifications.subscribe /^acfs\.operation/, collector2
     end
 
     describe '#process' do
       it 'should trigger event' do
         runner.process ::Acfs::Operation.new MyUser, :read, params: {id: 0}
         expect(collector.events).to have(1).items
+        expect(collector2.events).to have(1).items
       end
     end
 
@@ -59,14 +62,37 @@ describe ::Acfs::Runner do
       it 'should trigger event' do
         runner.run ::Acfs::Operation.new MyUser, :read, params: {id: 0}
         expect(collector.events).to have(1).items
+        expect(collector2.events).to have(0).items
       end
     end
 
     describe '#enqueue' do
       it 'should trigger event' do
-        runner.run ::Acfs::Operation.new MyUser, :read, params: {id: 0}
+        runner.enqueue ::Acfs::Operation.new MyUser, :read, params: {id: 0}
         expect(collector.events).to have(1).items
+        expect(collector2.events).to have(0).items
       end
+    end
+  end
+
+  describe '#run' do
+    before do
+      expect_any_instance_of(UserService).to receive(:prepare).and_return nil
+    end
+    it 'it should not do requests when a middleware aborted' do
+      expect(adapter).to_not receive :run
+      runner.run ::Acfs::Operation.new MyUser, :read, params: {id: 0}
+    end
+  end
+
+  describe '#enqueue' do
+    before do
+      expect_any_instance_of(UserService).to receive(:prepare).and_return nil
+    end
+    it 'it should not do requests when a middleware aborted' do
+      expect(adapter).to_not receive :queue
+      runner.enqueue ::Acfs::Operation.new MyUser, :read, params: {id: 0}
+      runner.start
     end
   end
 end